cc-6141 moved task creation business logic into task service

+++ src/Warranty.Core/Services/TaskService.cs
@@ -14,15 +14,30 @@ namespace Warranty.Core.Services
     {
         private IDatabase _database;
         private IEmployeeService _employeeService;
+        private IJobService _jobService;
 
-        public TaskService(IDatabase database, IEmployeeService employeeService)
+        public TaskService(IDatabase database, IEmployeeService employeeService, IJobService jobService)
         {
+            if (database == null)
+                throw new ArgumentNullException("database");
+            
+            if (employeeService == null)
+                throw new ArgumentNullException("employeeService");
+            
+            if (jobService == null)
+                throw new ArgumentNullException("jobService");
+
             _database = database;
             _employeeService = employeeService;
+            _jobService = jobService;
         }
 
         public bool TaskExists(Guid jobId, TaskType taskType)
         {
+            if (taskType == null)
+            {
+                throw new ArgumentNullException("taskType");
+            }
             using (_database)
             {
                 return _database.Single<int>(string.Format("SELECT COUNT(1) FROM Tasks WHERE ReferenceId = '{0}' AND TaskType = {1}", jobId, taskType.Value)) >= 1;
@@ -31,6 +46,10 @@ namespace Warranty.Core.Services
 
         public void CreateTaskUnlessExists(Guid jobId, TaskType taskType)
         {
+            if (taskType == null)
+            {
+                throw new ArgumentNullException("taskType");
+            }
             if (!TaskExists(jobId, taskType))
             {
                 CreateTask(jobId, taskType);
@@ -39,6 +58,10 @@ namespace Warranty.Core.Services
 
         public void DeleteTask(Guid jobId, TaskType taskType)
         {
+            if (taskType == null)
+            {
+                throw new ArgumentNullException("taskType");
+            }
             _database.Execute(string.Format("DELETE FROM Tasks WHERE ReferenceId = '{0}' AND TaskType = {1}", jobId, taskType.Value));
         }
 
@@ -47,9 +70,38 @@ namespace Warranty.Core.Services
             return _database.Fetch<Task>(string.Format("WHERE ReferenceId = '{0}'", jobId));
         }
 
-        public void CreateTask(Guid jobId, TaskType taskType)
+        private TaskType GetTaskTypeByStage(int stage)
         {
+            return TaskType.GetAll().SingleOrDefault(t => t.Stage.HasValue && t.Stage == stage);
+        }
 
+        public void CreateTasks(Guid jobId)
+        {
+            var job = _jobService.GetJobById(jobId);
+            if (_jobService.IsModelOrShowcase(job))
+            {
+                if(job.Stage == 7)
+                    CreateTaskUnlessExists(jobId, TaskType.JobStage7);
+            }
+            else 
+            {
+                if (job.Stage == 3 || job.Stage == 7 || job.Stage == 10)
+                {
+                    var taskType = GetTaskTypeByStage(job.Stage);
+                    if (taskType != null)
+                    {
+                        CreateTaskUnlessExists(jobId, taskType);
+                    }
+                }
+            }
+        }
+
+        private void CreateTask(Guid jobId, TaskType taskType)
+        {
+            if (taskType == null)
+            {
+                throw new ArgumentNullException("taskType");
+            }
             var wsr = _employeeService.GetWsrByJobId(jobId);
 
             using (_database)
